fix: get_active_session
