+++ src/koheesio/spark/utils/connect.py
@@ -4,26 +4,7 @@ from pyspark import sql
 from pyspark.errors import exceptions
 
 from koheesio.spark.utils import check_if_pyspark_connect_is_supported
-from koheesio.spark.utils.common import Column, DataFrame, ParseException, SparkSession
-
-
-def get_active_session() -> SparkSession:  # type: ignore
-    if check_if_pyspark_connect_is_supported():
-        from pyspark.sql.connect.session import SparkSession as ConnectSparkSession
-
-        session: SparkSession = (
-            ConnectSparkSession.getActiveSession() or sql.SparkSession.getActiveSession()  # type: ignore
-        )
-    else:
-        session = sql.SparkSession.getActiveSession()  # type: ignore
-
-    if not session:
-        raise RuntimeError(
-            "No active Spark session found. Please create a Spark session before using module connect_utils."
-            " Or perform local import of the module."
-        )
-
-    return session
+from koheesio.spark.utils.common import Column, DataFrame, ParseException, SparkSession, get_active_session
 
 
 def is_remote_session(spark: Optional[SparkSession] = None) -> bool:
@@ -54,14 +35,6 @@ def _get_spark_session_class() -> TypeAlias:
 
 def _get_parse_exception_class() -> TypeAlias:
     return exceptions.connect.ParseException if is_remote_session() else exceptions.captured.ParseException  # type: ignore
-
-
-# DataFrame: TypeAlias = _get_data_frame_class() if check_if_pyspark_connect_is_supported else sql.DataFrame  # type: ignore # noqa: F811
-# Column: TypeAlias = _get_column_class() if check_if_pyspark_connect_is_supported else sql.Column  # type: ignore # noqa: F811
-# SparkSession: TypeAlias = _get_spark_session_class() if check_if_pyspark_connect_is_supported else sql.SparkSession  # type: ignore # noqa: F811
-# ParseException: TypeAlias = (
-#     _get_parse_exception_class() if check_if_pyspark_connect_is_supported else exceptions.captured.ParseException  # type: ignore
-# )  # type: ignore  # noqa: F811
 
 
 __all__ = [
